+++ tests/testthat/test_read_flights.R
@@ -17,7 +17,7 @@ test_that("read_flights", {
   testthat::expect_equal( class(test1$nr_voo), 'numeric')
 
   # one month, combinada, no progress bar
-  test2 <- read_flights(date=202201, type='combinada' ,showProgress = FALSE)
+  test2 <- read_flights(date=202201, type='combinada', showProgress = FALSE)
   testthat::expect_true(is(test2, "data.table"))
   testthat::expect_true(nrow(test2) >0 )
 
@@ -25,7 +25,7 @@ test_that("read_flights", {
   testthat::expect_equal( as.character(min(test2$dt_referencia)), as.character("2022-01-01") )
 
   # all months in a year
-  test3 <- read_flights(date=2022, select=cols, showProgress = FALSE)
+  test3 <- read_flights(date=2022, select=cols, showProgress = FALSE, cache = TRUE)
   testthat::expect_true(is(test3, "data.table"))
   testthat::expect_true(nrow(test3) >0 )
 
@@ -36,6 +36,17 @@ test_that("read_flights", {
   months <- unique(test4$nr_mes_referencia)
   testthat::expect_true( all.equal(months , c(1,2)) )
 
+  # check whether cache argument is working
+  time_first <- system.time(
+    f201506 <- read_flights(date = 201506))
+
+  time_cache_true <- system.time(
+    f201506 <- read_flights(date = 201506, cache = TRUE))
+
+  time_cache_false <- system.time(
+    f201506 <- read_flights(date = 201506, cache = FALSE))
+
+  testthat::expect_true( time_cache_true[['elapsed']] < time_cache_false[['elapsed']] )
  })